[refactor] 파일 폴더 위치 이동
entity 폴더로 이동

#19
